requirements.txt
torch
torchvision
torchaudio
transformers==4.57.6
optimum==1.26.0
accelerate==1.8.0
peft==0.17.0
json-repair
Pillow
PyMuPDF
opencv-python-headless
numpy
tqdm
loguru